+++ Source/Core/ArcGisImageServerTerrainProvider.js
@@ -60,7 +60,7 @@ define([
      *   token : 'KED1aF_I4UzXOHy3BnhwyBHU4l5oY6rO6walkmHoYqGp4XyIWUd5YZUC1ZrLAzvV40pR6gBXQayh0eFA8m6vPg..',
      *   proxy : new Cesium.DefaultProxy('/terrain/')
      * });
-     * scene.terrainProvider = terrainProvider;
+     * viewer.terrainProvider = terrainProvider;
      */
     var ArcGisImageServerTerrainProvider = function ArcGisImageServerTerrainProvider(options) {
         //>>includeStart('debug', pragmas.debug);